# src/main/resources/api.yaml
openapi: 3.0.0
info:
   version: 1.0.0
   title: Vehicles
   description: Polymorphism example
paths:
   /vehicles:
      get:
         summary: List all vehicles
         tags:
         - vehicles
         responses:
            '200':
               description: An paged array of vehicles
               content:
                  application/json:
                     schema:
                        type: array
                        items:
                           $ref: '#/components/schemas/Vehicle'
            default:
               description: unexpected error
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Error'
   /vehicles/{id}:
      get:
         summary: Info for a specific vehicle
         tags:
         - vehicles
         parameters:
         -  name: id
            in: path
            required: true
            description: The id of vehicle to retrieve
            schema:
               type: string
         responses:
            '200':
               description: Expected response to a valid request
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Vehicle'
            default:
               description: unexpected error
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Error'
components:
   schemas:
      Vehicle:
         type: object
         properties:
            id:
               type: integer
            type:
               type: string
               enum:
               - CAR
               - PLANE
            model:
               type: string
            name:
               type: string
            buildDate:
               type: string
               format: date-time
         discriminator:
            propertyName: type
            mapping:
               CAR: '#/components/schemas/Car'
               PLANE: '#/components/schemas/Plane'
         required:
         - id
         - type
      Car:
         allOf:
         -  $ref: '#/components/schemas/Vehicle'
         -  type: object
            properties:
               has_4_wheel_drive:
                  type: boolean
      Plane:
         allOf:
         -  $ref: '#/components/schemas/Vehicle'
         -  type: object
            properties:
               has_reactor:
                  type: boolean
               nb_passengers:
                  type: integer
      Error:
         required:
         - code
         - message
         properties:
            code:
               type: integer
               format: int32
            message:
               type: string